Add fuzzy patch exception
As written, made it impossible to do fuzzy patches. So add an exception for covered users to legally include portions of code but only the minimum required to patch. Also define patch more clearly.
Add license survival clause so that estates may exercise authority as well. We're not getting any younger here.